+++ tests/acceptance/bootstrap/FeatureContext.php
@@ -148,7 +148,6 @@ class FeatureContext extends BehatVariablesContext {
 	private int $ocsApiVersion = 1;
 	private ?ResponseInterface $response = null;
 	private string $responseUser = '';
-	public array $emailRecipients = [];
 	private CookieJar $cookieJar;
 	private string $requestToken;
 	private array $createdFiles = [];
@@ -278,18 +277,6 @@ class FeatureContext extends BehatVariablesContext {
 		} catch (Exception $exception) {
 			// if response couldn't be converted into xml then push "notset" to last ocs status codes array
 			$this->pushToLastOcsCodesArray("notset");
-		}
-	}
-
-	/**
-	 * @param string $emailAddress
-	 *
-	 * @return void
-	 */
-	public function pushEmailRecipientAsMailBox(string $emailAddress): void {
-		$mailBox = explode("@", $emailAddress)[0];
-		if (!\in_array($mailBox, $this->emailRecipients)) {
-			$this->emailRecipients[] = $mailBox;
 		}
 	}
